requirements.txt
Flask
python-dotenv
langchain
python-docx
langchain-groq          # for ChatGroq integration